app/requirements.txt
requests
flask
pyyaml